Move exceptions into the Format module to make names shorter
They're distinguishable enough for this to be okay and it makes sense
to put them here because this is where we will deal with them (as they
propagate up to here).
