+++ lib/jerboa/format/header.ex
@@ -5,22 +5,6 @@ defmodule Jerboa.Format.Header do
 
   @magic_cookie MagicCookie.encode
 
-  defmodule First2BitsError do
-    defexception [:message, :bits]
-
-    def message(%__MODULE__{}) do
-      "the most significant two bits of a STUN message must be zeros"
-    end
-  end
-
-  defmodule MagicCookieError do
-    defexception [:message, :header]
-
-    def message(%__MODULE__{}) do
-      "STUN message doesn't have magic cookie"
-    end
-  end
-
   def encode(params) do
     t = Type.encode(params)
     l = Length.encode(params)
@@ -38,10 +22,10 @@ defmodule Jerboa.Format.Header do
     end
   end
   def decode(%Jerboa.Format{header: <<0::2, _::30, _::128>> = header}) do
-    {:error, MagicCookieError.exception(header: header)}
+    {:error, Jerboa.Format.MagicCookieError.exception(header: header)}
   end
   def decode(%Jerboa.Format{header: <<b::2, _::158>>}) do
-    {:error, First2BitsError.exception(bits: b)}
+    {:error, Jerboa.Format.First2BitsError.exception(bits: b)}
   end
 
   defp encode(type, length, identifier) do
