+++ lib/jerboa/format/header/length.ex
@@ -1,13 +1,6 @@
 defmodule Jerboa.Format.Header.Length do
   @moduledoc false
-
-  defmodule Last2BitsError do
-    defexception [:message, :length]
-
-    def message(%__MODULE__{}) do
-      "all STUN attributes are padded to a multiple of 4 bytes so the last 2 bits of this field should be zero"
-    end
-  end
+  alias Jerboa.Format.Last2BitsError
 
   def encode(%Jerboa.Format{body: b}) when is_binary(b) do
     <<byte_size(b)::integer-unit(8)-size(2)>>
