+++ test/jerboa/format/body_test.exs
@@ -1,5 +1,6 @@
 defmodule Jerboa.Format.BodyTest do
   use ExUnit.Case, async: true
+  alias Jerboa.Format
   alias Jerboa.Format.Body
   alias Jerboa.Format.Body.Attribute
 
@@ -36,7 +37,7 @@ defmodule Jerboa.Format.BodyTest do
 
     test "unknow comprehension required attribute results in :error tuple" do
       for x <- 0x0000..0x7FFF, not x in known() do
-        assert {:error, %Attribute.ComprehensionError{attribute: x}} == Attribute.decode(%Jerboa.Format{}, x, <<>>)
+        assert {:error, %Format.ComprehensionError{attribute: x}} == Attribute.decode(%Jerboa.Format{}, x, <<>>)
       end
     end
   end
